+++ src/main/java/com/spring/test/SpringSecurity_JWT_test/repository/UserRepository.java
@@ -7,12 +7,12 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 //@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
 
     Optional<User> findByUsername(String username);
     User findByEmail(String email);
     Boolean existsByUsername(String username);
     Boolean existsByEmail(String email);
 
-     public Optional<User> findById(Long id);
+     public Optional<User> findById(Integer id);
 }
